multi-level-vector-index/requirements.txt
numpy
faiss-cpu
h5py
requests
matplotlib
pytest